requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
djangorestframework==3.13.1
heroku==0.1.4
idna==3.3
psycopg2==2.9.3
python-dateutil==1.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
